update proto.lock
In this change we made breaking changes, which is fine for now. However, in preparation for post-production release of blueprint, we've added checks for this. To ensure checks pass the proto.lock file has to be update to date. We can commit breaking changes to the proto.lock by using the commands highlighted in the readme.